Update same_query_across_sites example
