+++ examples/compare_product_prices/README.md
@@ -1,0 +1,16 @@
+# Example script: comparing product price across websites with AgentQL
+
+This example demonstrates how to compare product prices across websites with query_data() method.
+
+## Run the script
+
+* [Install AgentQL SDK](https://docs.agentql.com/docs/installation/sdk-installation)
+* Save this python file locally as **compare_product_prices.py**
+* Run the following command from the project's folder:
+```bash
+python3 compare_product_prices.py
+```
+
+## Play with it
+
+Install the [AgentQL Chrome DevTools extension](https://docs.agentql.com/docs/installation/chrome-extension-installation/) to play with the AgentQL query. [Learn more about the AgentQL query language](https://docs.agentql.com/docs/agentql-query/query-intro)
